\name{show.sim.weed}
\alias{show.sim.weed}

\title{ Plot simulated weed dataset }
\description{Takes the output of \code{sim.weed} and plot the different
  variables according to the options of simulations}
\usage{
show.sim.weed(sim) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sim}{ A list as  produced by \code{sim.weed}}
}

\value{No value is returned}

\references{ G. Guillot, N. Loren, M. Rudemo, Bayesian spatial prediction of weed intensities from  exact count 
  data  and picture based indexes, 2006, submitted}
\author{Gilles Guillot}



\keyword{}

